+++ documents/forms.py
@@ -3,7 +3,7 @@ from django import forms
 
 class RejectionCommentForm(forms.Form):
     rejection_comment = forms.CharField(
-        required=False,  # Делаем поле необязательным
+        required=False,
         widget=forms.Textarea(attrs={'rows': 24}),
         label='Комментарий к отклонению'
     )
